Modified post processing so that it can process more than 256 configurations

+++ scripts/log_processing/processConfig.py
@@ -193,8 +193,8 @@ def process_config(base_folder, config, sample_rate):
 
     # get parameters
     base_folder = os.path.join(base_folder, '')
-    config_folder = os.path.join(base_folder,'configs', config, '')
-    config_number = np.uint8(config[1:])  # drop letter c and parse number
+    config_folder = os.path.join(base_folder, 'configs', config, '')
+    config_number = np.uint16(config[1:])  # drop letter c and parse number
     step_size = np.float32(0.08)
 
     # get maze metrics:
